refactor yuidoc theme to take advantage of @element
calls elements Components
components will be documented as @element
in the case of editor widgets, their functionality
is provided as mixins (to allow the user the ability
to compose widgets) and the mixins documented as classes.
In the case of  and  there
will simply be two yuidoc blocks, one with @class and one
with @element. In the future I may look at hacking yuidoc
so that I can place the @element docs in the XXX.hbs file,
or at least so that for components the Defined in: ... points
to the .hbs file.